Split test_framework.py (1873 LoC) into separate files
